I was half asleep when debugging.  I made the problem worse where it was adding Xdelta and Ydelta to the index of the register.  It should have been used to delta the number in the register.
However there is still a bug.  Starts drawing a sprite 8xN at 63x12.  In the second loop of the horizontal draw that leads to it trying to draw beyond the screen 64x12 (max is 63).  Unsure whether it's a fault in an earlier part of code leading to bad data or hardware expects it to wrap to other side of screen.